+++ CommonCPPCTest.cmake
@@ -2,14 +2,17 @@
 #               2010-2015, Stefan Eilemann <eile@eyescale.ch>
 #               2014, Juan Hernando <jhernando@fi.upm.es>
 #
-# Creates cpptests and perftests targets. The first one creates a test
-# for each .c or .cpp file in the current directory tree, excluding
-# the ones which start with perf (either as top-level directory name
-# or filename). The latter are added to perftests, and are not meant
-# to be executed with each test run. Both targets will update and
-# compile the test executables before invoking ctest. CommonCTests
-# includes the targets into the more general 'tests' and 'perftests'
-# targets, respectively.
+# Creates per project cpptests, perftests and nightlytests targets.
+# The first one creates a test for each .c or .cpp file in the current
+# directory tree, excluding the ones which start with perf or nightly (either
+# as top-level directory name or filename).
+# For each file starting with perf or nightly a test is created and added to
+# pertests and nightlytests respectively. These two targets are not meant
+# to be executed with each test run. All targets will update and compile the
+# test executables before invoking ctest. The nightlytests target will also
+# run pertests.
+# The per project targets are included into the global 'tests', 'perftests' and
+# 'nightlytests' from CommonCTest.cmake.
 #
 # Input:
 # * TEST_LIBRARIES link each test executables against these libraries
@@ -45,8 +48,9 @@ foreach(FILE ${EXCLUDE_FROM_TESTS})
 endforeach()
 list(SORT TEST_FILES)
 
-set(ALL_CPP_TESTS)
+set(ALL_CPP_UNIT_TESTS)
 set(ALL_CPP_PERF_TESTS)
+set(ALL_CPP_NIGHTLY_TESTS)
 
 # backwards compat: generate main() for unit tests
 #  should really #define BOOST_TEST_DYN_LINK if using boost
@@ -100,8 +104,11 @@ macro(common_add_cpp_test NAME FILE)
     set(TEST_LABELS ${PROJECT_NAME}-perf)
     install(TARGETS ${TEST_NAME} DESTINATION share/${PROJECT_NAME}/benchmarks
       COMPONENT apps)
+  elseif("${NAME}" MATCHES "^nightly" OR "${NAME}" MATCHES "nightly-")
+    list(APPEND ALL_CPP_NIGHTLY_TESTS ${TEST_NAME})
+    set(TEST_LABELS ${PROJECT_NAME}-nightly)
   else()
-    list(APPEND ALL_CPP_TESTS ${TEST_NAME})
+    list(APPEND ALL_CPP_UNIT_TESTS ${TEST_NAME})
     set(TEST_LABELS ${PROJECT_NAME}-unit)
   endif()
 
@@ -136,35 +143,31 @@ if(CMAKE_VERSION VERSION_GREATER 3.1.99)
   set(__CONSOLE USES_TERMINAL)
 endif()
 
-if(NOT TARGET ${PROJECT_NAME}-cpptests)
-  add_custom_target(${PROJECT_NAME}-cpptests ${__CONSOLE}
-    COMMAND ${CMAKE_CTEST_COMMAND} -T test --no-compress-output
-    --output-on-failure -L ${PROJECT_NAME}-unit -C $<CONFIGURATION> \${ARGS}
-    WORKING_DIRECTORY ${PROJECT_BINARY_DIR}
-    COMMENT "Running ${PROJECT_NAME} unit tests")
-endif()
-set_target_properties(${PROJECT_NAME}-cpptests PROPERTIES
-  EXCLUDE_FROM_DEFAULT_BUILD ON FOLDER ${PROJECT_NAME}/tests)
+macro(ADD_TEST_TARGET display_name target_part label_part)
+  if(NOT TARGET ${PROJECT_NAME}-${target_part}tests)
+    add_custom_target(${PROJECT_NAME}-${target_part}tests ${__CONSOLE}
+      COMMAND ${CMAKE_CTEST_COMMAND} -T test --no-compress-output
+      --output-on-failure -L ${PROJECT_NAME}-${label_part} -C $<CONFIGURATION> \${ARGS}
+      WORKING_DIRECTORY ${PROJECT_BINARY_DIR}
+      COMMENT "Running ${PROJECT_NAME} ${display_name} tests")
+  endif()
+  set_target_properties(${PROJECT_NAME}-${target_part}tests PROPERTIES
+    EXCLUDE_FROM_DEFAULT_BUILD ON FOLDER ${PROJECT_NAME}/tests)
+  string(TOUPPER ${label_part} LABEL_PART)
+  if(ALL_CPP_${LABEL_PART}_TESTS)
+    add_dependencies(${PROJECT_NAME}-${target_part}tests
+                     ${ALL_CPP_${LABEL_PART}_TESTS})
+  endif()
+  add_dependencies(${target_part}tests ${PROJECT_NAME}-${target_part}tests)
+endmacro()
 
-if(NOT TARGET ${PROJECT_NAME}-perftests)
-  add_custom_target(${PROJECT_NAME}-perftests ${__CONSOLE}
-    COMMAND ${CMAKE_CTEST_COMMAND} -T test --no-compress-output
-    --output-on-failure -L ${PROJECT_NAME}-perf -C $<CONFIGURATION> \${ARGS}
-    WORKING_DIRECTORY ${PROJECT_BINARY_DIR}
-    COMMENT "Running ${PROJECT_NAME} performance tests")
-endif()
-set_target_properties(${PROJECT_NAME}-perftests PROPERTIES
-  EXCLUDE_FROM_DEFAULT_BUILD ON FOLDER ${PROJECT_NAME}/tests)
-
-add_dependencies(${PROJECT_NAME}-cpptests ${ALL_CPP_TESTS})
-if(ALL_CPP_PERF_TESTS)
-  add_dependencies(${PROJECT_NAME}-perftests ${ALL_CPP_PERF_TESTS})
-endif()
+add_test_target(unit cpp unit)
+add_test_target(performance perf perf)
+add_test_target(nightly nightly nightly)
 
 add_dependencies(${PROJECT_NAME}-tests ${PROJECT_NAME}-cpptests)
 add_dependencies(tests ${PROJECT_NAME}-tests)
-add_dependencies(cpptests ${PROJECT_NAME}-cpptests)
-add_dependencies(perftests ${PROJECT_NAME}-perftests)
+add_dependencies(${PROJECT_NAME}-nightlytests ${PROJECT_NAME}-perftests)
 
 if(ENABLE_COVERAGE)
   add_coverage_targets(${PROJECT_NAME}-cpptests)